feat: refactor signup components and add new input fields with improved structure
