+++ web/src/app/globals.css
@@ -10,6 +10,7 @@
   --color-foreground: var(--foreground);
   --color-ssa-red: #f85b76;
   --color-ssa-red-light: #ff879c;
+  --color-ssa-red-pale: #ffb3bf;
   --color-ssa-yellow: #ffe6b6;
   --color-ssa-yellow-light: #fff7e9;
   --color-ssa-white: #fefcf9;
