+++ web/src/components/PaymentStep.tsx
@@ -15,8 +15,7 @@ export default function PaymentStep({
         </p>
         <ul className="list-disc list-inside text-sm text-ssa-black space-y-1 ml-1">
           <li>
-            Goodies and discounts from SSA sponsors when you show them your
-            membership sticker
+            Goodies and discounts from SSA sponsors when you show them your membership sticker
           </li>
           <li>Please be sure to collect your MEMBERSHIP CARD from the team.</li>
         </ul>
@@ -24,8 +23,7 @@ export default function PaymentStep({
           <button
             onClick={onPay}
             disabled={isLoading}
-            className="w-full py-2.5 rounded-lg text-white text-sm font-medium disabled:opacity-60"
-            style={{ backgroundColor: '#635bff' }}
+            className="w-full py-2.5 rounded-lg bg-[#635bff] text-white text-sm font-medium disabled:opacity-60"
           >
             {isLoading ? 'Processing...' : 'Pay'}
           </button>
